Reorganize src/test files for module aliasing Resolves rdar://93611579
